+++ changelog.md
@@ -1,6 +1,6 @@
 ### 1.3.0 (W.I.P.)
 - Add commands to print TFC's level settings and chunk data
-- Add setting (default enabled) to deduplicate TFc datapack validation errors that are inserted into the KubeJS console
+- Add setting (default enabled) to deduplicate TFC datapack validation errors that are inserted into the KubeJS console
 - Add chunk generator type `kubejs_tfc:wrapped` which enables adding chunk data to levels without proper TFC-like level generators
 - New event for creating chunk data, depends on above chunk generator type
 - De-hardcode dead crop model generation
@@ -11,7 +11,7 @@
   - `#drips` in `tfc:thin_spike` block type, same functionality handled by nullable parameter of `#dripParticle`
   - `#hasBubbles` & `#hasSteam` in `tfc:spring` fluid type, same functionality handled by nullable parameter of `#bubbleParticle` & `#steamParticle`
   - `#allModels` in berry bushes, superseded by `#models`
-- Remove deprecated methods in custom climate model event
+  - Deprecated methods in custom climate model event
 - Remove old, non-functional chunk-data creation mechanism for custom climate models that was added in 1.2.6
 - Add support for Kube's ingredient and result modifiers to TFC's advanced crafting recipes
 - Fix drinkable description command printing the first effect in place of all-but-the-last effect